pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dns-hygiene-guard"
version = "0.1.0"
description = "Automation tooling to audit DNS hygiene for self-managed zones."
authors = [{ name = "DNS Hygiene Guard Maintainers" }]
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "dnspython>=2.4",
    "PyYAML>=6.0",
    "idna>=3.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
dng = "dns_hygiene_guard.__main__:main"

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = [
    "tests",
]


[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"dns_hygiene_guard" = ["demo/data/**/*"]

[tool.ruff.lint.isort]
known-first-party = ["dns_hygiene_guard"]
